Mod: CodeClimate clean code
